Allow track points to be moved by dragging
In support of this, draw the track/point graphics and their
dragging animation on a separate Marble rendering layer.  This
doesn't actually help with eliminating the need to continuously
repaint the complex map graphics while a drag is going on, but
it does separate out our application-specific painting and
there is always a chance that Marble will do this in future.