Make the tracks and waypoints display layers into subclasses
New common base class LayerBase which handles most of the painting,
selection and dragging.  It uses classification functions provided
by the specific layer class to select what types of item and over
which parts of the tree it operates.  The lowest level of painting
is handled by the layer class.