Big rewrite for tracking of data items in commands.
The reference counting system was clumsy.  Enforce a better
solution by mandating that a TrackDataItem can only have one
parent (or none) at any particular time.  A command can only
retain a simple pointer to an item if it is not going to
manipulate it in any way other than in place;  otherwise,
it must hold it in a container.  The command need then only
delete the container in its destructor, and any items that
it holds will be deleted without having to worry about whether
they are also needed as part of the main data tree.

Do all of the item manipulation directly within the commands
themselves, so eliminating many of the functions within the model.
This is actually more efficient because the model change signals
can be emitted only once even if the command needs a loop.

Lazy allocate the item's list of children and metadata.