Waypoint import/export: Support for folders
A folders is not actually a real entity within the GPX file.
Instead, the folder path is stored (as a private application
extension) with each waypoint, and created if necessary upon
import.

The only drawback with this is that empty folders will not
be stored in the file or therefore created.