Add a user-selectable status for notes/photos
To keep things simple, this can be applied to any waypoint.
The status can be none, To Do or Done, and the waypoint name
is shown in colour if a status is set.

There is an additional "Status" combo box on the properties
"General" page.  To keep things simple, this can be applied to
any waypoint.  It - and currently only this data - can also be
applied to multiple waypoints.