ProfileWidget: Correct the display of X axis times
It is not possible to make QCustomPlot display them as time zone times
(the time zone of the file), which is what it required here.  So the
time zone offset is adjusted explicitly and the QCustomPlot is told
that they are UTC.

Update the plot on a timer, to avoid multiple updates if more that one
control changes (e.g. one radio button going on and another one off).