GCC 5.4: Correct use of isnan()
Regardless of whether the macro is defined as std::isnan() or isnan(),
the compilation fails on certain source files (notably trackdata.cpp and
filescontroller.cpp will work with one setting, all other files need
another).  Doing such a wide-ranging thing as 'using namespace std' in a
header file is not normally a good thing, but in this case it is
necessary.