Fix a selection of both tracks and folders not being considered "mixed"
This is because the check in FilesView::selectionChanged() tests
for a "mixed" selection by checking whether all selected items have
the same parent.  In this case they do and so the selection is not
considered to be "mixed".

To cover this case it also needs to look at the item types.  Add a
virtual type() function to all track data items to eliminate the long
series of dynamic_cast's formerly used here.