Do not save automatically-assigned names in the GPX file
They are needed within this application, but there is no need for them
to be saved in the GPX file - which expands its size and invents
information that was not there before.  Instead, keep track of whether
an item's name has been set explicitly (i.e. as a user operation) and
only export the name if this is the case.

This allows the constructors of TrackDataItem and everything derived
from it to be cleaned up, eliminating the name parameter.