Properties dialogue: Hide the "Style" page if not applicable
To the item type, for example for a trackpoint.  It is still present but
disabled if not applicable to multiple items.

This has a number of implications.  The tab handling and validation is
now not specific to the presence or ordering of any tabs.  Also the
last used dialogue tab is saved separately for the item type, because a
global setting may not be valid for the next dialogue for a different
type.

It does mean, though, that an item type can indicate that it has no
style page by returning a null pointer instead of constructing the page.
This is done by using the NULL_PROPERTIES_PAGE macro as opposed to the
usual CREATE_PROPERTIES_PAGE;  no constructor is then required.

The intention of all of this is to simplify future expansion with
other pages, e.g. a "Plot" page (bearing lines and range rings) for
a waypoint.