ProfileWidget: Show waypoints on plot
First implementation.  Find the plot point which is closest to the
waypoint, subject to a minimum distance, then plot the waypoint icon and
name on the elevation graph at the corresponding point.  Choosing the
elevation graph because that is the most likely to have useful data, and
it is a physical property matching to a physical location.