Start the conversion of point metadata QString -> QVariant
For a number of reasons.  Firstly, so that the metadata can be used for
stroring numeric or date/time values without having to be converted to
string form, eliminating conversions and potential loss of precision
as well as better efficiency.  This will allow some data (such as elevation
and time) to be consistently handled as metadata instead of being a special
case.  Secondly, to allow the metadata to be edited directly, using Qt's
built in editing support for item views.