Eliminate the Style class and data
It was originally intended to be a way of collecting together and
organising style-related information for all types of items.  However,
it turned out to be a less elegant solution than expected, with not
very much information to organise - only the line and point colours -
which can now be easily handled as part of the QVariant-based metadata.
So the style.* source files and the Style class are completely removed.