Start of big rewrite of Properties dialogue
To store all the data being managed as metadata.  This is so that there
will be a single consistent data store, with any changes automatically
being reflected to all parts of the dialogue including the metadata
table.  Which will eliminate the need for special hacks when, for
example, the time zone or the point position needs to be changed,
although the time zone is still a special case because it affects the
display not only of its own GUI widget but also all displayed times.  It
will also much simplify the individual cases for each type of data in
FilesController::slotTrackProperties(), with only a small and limited
number of the properties needing special handling while the others can
be handled in a loop over all metadata.

This commit implements the base data store, the dialogue GUI framework
and the "General" page.