TrackDataItem::timeZone():  Do not be file-specific
That is, where there is "timezone" metadata at whetever level in the
ancestor tree, including at the level of an item itself, then accept it.
This allows the possibility for separate tracks, for example, to have
different time zones.  However, there is no GUI for setting a time zone
on anything other than a top level file.