More intelligent "Expand"/"Collapse" operations for the tree view
Collapsing everything to just the root file item is not useful.  Neither
is expanding everything including long segments.  So collapse down to
the second level items (tracks, folders or routes) and expand everything
apart from segments and routes.  Keep the old actions with a different
shortcut and not shown in the menus.