Rename action "Move Points" to "Move Mode" and fix for routes
The start of using this action to also enable dragging and dropping
in the tree view.  Because the conditions to be allowed to do it, and
the conditions for turning it off, are the same.

Do not enable the mode for folders (which may contain waypoints, but a
waypoint needs to be selected to enable the mode), but do enable it for
a route.