Allow the file time zone to be changed even if it is read only
So that correct timews can be examined.  Set via a new menu action
"Data - Set Time Zone" which is more discoverable than setting the
properties of the root file item (the old way still works).  Allow
this change to be made in the properties dialogue and take effect
even if the file is read only.