Implement application-independent application data storage
Yes, that sounds paradoxical.  What it really means is to move the
global application data previously held by MainWindow into a new
ApplicationData class, which only stores and refers to the MainWindow
and its FilesController and MapController by pointers.  This is so that
the ApplicationData can be part of 'core' and it can be accessed by
any part of the application source.

The MainWindowInterface is renamed to ApplicationDataInterface, which
searches up the Qt object tree for the ApplicationData as before.

The data previously held by the Project class is also moved into the
ApplicationData in order to keep everything together in one place.