Make the FilesView accessible via ApplicationDataInterface
This, along with removing the dependency of MapController on
FilesController (emit a signal instead when dragging points on the map),
enables everything to do with the map to be moved from 'app' to a new
subdirectory 'map'.

Add a mainWidget() to the ApplicationDataInterface so that it can be
used as a QWidget (e.g. for a widget parent) without having to have the
definition of MainWindow available.  This may allow other components to
be moved out of 'app' later.