DataIndexer: Support for XML namespaces
Associate a tag name with a namespace, either preset for those tags that
we know are private to this application or as found in the imported GPX
file for others.  This centralises all knowledge of which tag names are
private to this application, and the application namespace prefix,
in DataIndexer.