When creating a route point, take container from selected route point
If there is one.  This allows a route to be quickly created without
having to select the destination route every time in the dialogue.

The choice of container is done in FilesController::slotAddRoutepoint(),
because this has better knowledge of what is allowed and appropriate
than the generic dialogue.  CreatePointDialogue::setSelectedContainer()
therefore actually does something now.  The previous logic in the
CreatePointDialogue constructor is kept as a fallback.