Working models for the points list view
Not sure why the setup

  view -> QSortFilterProxyModel -> KDescendantsProxyModel -> FilesModel
              (PointsModel)

didn't work - the data was displayed, but nothing could be selected in
the view.  The same happened if the order of the two intermediate models
was reversed.

Build an internal list of applicable points in PointsModel, and use it
to access the data items as the source of the model data.