Allow some properties of multiple items to be changed
Not trivial, because FilesController::slotTrackProperties() assumes
that the data returned from the dialogue is either unchanged from
or is to be applied to the selected items, based on the original
value from the first item only.  Instead of assuming this, note
which data has been changed in MetadataModel and use that information
to update the selected items just with the changed data, leaving any
unchanged data alone even if there are multiple values within the
selected items.

Data that has been changed is identified on the "Metadata" tab by
colour.

Some values may still not be able to be changed for multiple items,
because they are disabled in the GUI.  This includes the "Description"
field and the "Plot" tab.