Waypoint icons: Provision for named symbols
Rename WaypointImageProvider -> PointIconProvider for obtaining
icons for all types of points, not just waypoints.  Caches
results for efficiency.

New PointIcon to extend QIcon with a name, for caching and
display.  Can be constructed from either a name (in an optional
namespace, allowing support for icons from external applications)
or a waypoint color.  Can derive a QIcon or a QPixmap when needed.

Display and sort by symbol names in PointsModel.

Support for symbol names (the SYM tag) in TrackDataWaypoint::icon(),
with a priority order for the icon source.