Make category list file-specific
With a simplified allocatable list class CategoriesList (renamed
from CategoriesManager) which is attached to the TrackDataFile root
item.  A waypoint can search upwards from itself to find the root and
therefore access the category list.  This actually makes loading or
importing a file simpler, because the ImportFileCommand can do the
merging as a simple add without overwrite.

Still to do: The import and merge of categories is not completely
undo'able.