GPX exporter: Big rewrite for correct container structure
In order to organise the various containers that may or may not
be required by GPX and consuming applications.  Collect together
all of the data to be output in a series of queues, then output
each enclosed in appropriate container tags.  Eliminates
startExtensions()/endExtensions() and keeping track of whether
the extensions container has been started.