GPX importer: Do not set line/point colour from COLOR data
If this were done, then it would be permanently recorded in the file
if it were to be reexported.  Instead, record both our application
colours and the fallback colour separately, if defined for an imported
item.  Fall back to the colour where required - when generating a
waypoint icon or when resolving a map line colour.