Add Waypoint: Create a folder to contain it if none already exists
For use in points list mode with an empty file, where it is not
obvious from the GUI that a folder needs to be created first (and,
even if the user does know this, needs a switch to tree view and
then back again).

Implement AddContainerCommand::addedItem() for use when creating
folder as above, which eliminates a static variable.  Also use this
to create the folder during stops detection.  Also implement similar
AddWaypointCommand::addedItem() which can be used for photo import
again without the use of the static variable.

Because of this implementation. the folder creation in the 3 cases
above will appear separately in the undo history.