ProfileWidget: Do not assert if no input points are available
The action should normally be disabled by the GUI, but due to the
full rigourous checks being too expensive it is assumed that a
selected file, track or route contains some valid points.  This
may not be true, so tell the user if there are no points available
when displaying the profile window.

The same applies to stop detection, it does not assert but there
will be no meaningful results.