Extend category storage and management to include OsmAnd data
That application can specify an icon name and background shape
for categories (which it calls "groups"), not just a colour.
So make a category into a opaque CategoryData type holding the
colour, icon and shape, with the category list mapping from the
name to that.

Change source file and type names to better reflect their function.
Update categories import, export and management (currently only
using colour as before) to reflect those changes.