Reorganise data models using Qt/Frameworks provided ones wherever possible
In particular, for those that feed the points list view.  Instead of
PointsModel generating a private list of waypoints by recursively
scanning the main data tree, flatten the tree and then filter the
list using a chain of models.  PointsModel can then just use a
KExtraColumnsProxyModel to add extra columns to that list to display
the appropriate data.

Use a KLinkItemSelectionModel to synchronise the selection between
the tree view and the points list view.  This removes a number of
functions and signals in both FilesModel and PointsModel.

Rename FilesController::model() to filesModel() to better reflect its
function.