ExportFileDialogue: Use Qt/Frameworks models for the home points list
Starting with the flattened waypoint list model as added in 7ad6e299,
filter it for home waypoints only and then add display data to it.

Eliminate the "None" line from the list, no point can be specified
simply by setting all of the check boxes in the column to off.