Use a small separate dialogue for "Set Time Zone..."
Setting the time zone by selecting the top level file item and
entering the properties dialogue was always a hack, with lots of
special accomodation for the case where the file was read only.
Use a separate dialogue instead and immediately modify the file
in memory if it is read only; otherwise, modify the file via an
undoable command in the same way as any other item data change.