Implement ItemIndexInterface::rootItem() for consistent access
Deprecate the former FilesModel::rootFileItem() which returns a
TrackDataFile pointer - rootItem() returns a TrackDataItem.  This
allows the former to be deprecated (functions can't be overloaded
by return type alone) and would allow a possible return to the former
"project" structure where the data tree root may not be a file but
a collection of them.