Do not treat Garmin SYM and OsmAnd ICON as equivalent
So as to not lose a value set for one because a symbol is set
for the other.

Implement PointIcon::metadataKey() and AbstractIconProvider::metadataKey()
so that a provider can indicate which tag its data is expected to
be stored under.

Implement TrackDataWaypoint::createPointIcon() to help with icon creation.

Prioritise icon name lookup in TrackDataWaypoint::icon() to first use
any explicitly named SYMSET, then look for OsmAnd and Garmin icons in
that order.