OsmAnd icons: Setting for the icons source location
Rather involved, because the icon provider (in 'core') does not have
access to the application settings directly.  Therefore SettingsDialogue
uses PointIcon::setProviderOption() to pass a named option and value to
a specified provider, which in turn calls AbstractIconProvider::setOption()
for the appropriate provider.  This happens at application startup and
when the settings have been changed.

Although there is only one provider option needed at the moment, at least
this approach is extensible.