Properties pages: Show a warning if file time zone is not set
Which means that all the displayed times will be in UTC - this is
shown anyway by them not having a time zone suffix, but the message
makes it clear.  Less obtrusive than always checking the time zone
when loading a file.

The warning is only shown on the "General" and "Details" pages
because they are the only ones that show times.