MediaPlayer: Do not insist that the TrackDataItem is a waypoint
Whn called from TrackWaypointDetailPage to populate the "Media"
field, the 'item' passed is the internal item from MetadataModel
which is really a TrackDataTrackpoint.  Therefore the AS() at the
beginning of findMediaFileInternal() will fail and a null URL is
returned.  Ths causes the media link field to be displayed as blank.

Since the TrackDataWaypoint is only used to check the actual media
type against what is expected, and TrackWaypointDetailPage calls
findMediaFile() any media type is allowed, so there is no need to
perform the test if the cast fails.

Rename the static findMediaFile() -> findMediaFileInternal() to
reduce confusion.